Fix/Add: Add password on sign up and fix register file with partial html page
