+++ src/Pages/Register/Register.scss
@@ -49,35 +49,10 @@
                 font-weight: 700;
             }
         }
-        .content__registerName {
-            display: flex;
-            justify-content: center;
-            position: relative;
-            padding: 25px 0 0 0;
-
-            @media(max-width: $TabletM) {
-                padding: 25px 25px 0 25px;
-            }
-
-            input[type="text"] {
-                background: transparent;
-                border: transparent;
-                width: 536px;
-                height: 60px;
-                font-size: 16px;
-                border: 1px solid $border_color_light;
-                border-radius: 5px;
-                color: white;
-            
-                @media(max-width: $TabletM) {
-                    width: 100%;
-                }
-            }
-        }
         .content__registerName:focus {
             border: 1px solid $blue_Twitter;
         }
-        .content__registerEmail {
+        .content__register {
             display: flex;
             justify-content: center;
             position: relative;
@@ -87,7 +62,7 @@
                 padding: 25px 25px 0 25px;
             }
 
-            input[type="text"] {
+            input[type="text"], input[type="password"] {
                 background: transparent;
                 border: transparent;
                 width: 536px;
@@ -106,7 +81,7 @@
             border: 1px solid $blue_Twitter;
         }
         .content__changeEmailPhone{
-            padding: 30px 0 0 33px;
+            padding: 20px 0 0 33px;
             .registerChangeEmailPhone{
                 cursor: pointer;
                 background: transparent;
@@ -118,7 +93,7 @@
             }
         }
         .content__dateInfo{
-            padding: 35px 0 0 33px;
+            padding: 20px 0 0 33px;
             .Birthday{
                 font-size: 15px;
                 font-weight: 700;
@@ -138,7 +113,7 @@
             }
         }
         .content__birthdaySelect{
-            padding: 16px 25px 30px 33px;
+            padding: 16px 25px 10px 33px;
             display: flex;
             width: 100%;
             align-items: center;
